+++ src/main/java/com/commercetools/sync/producttypes/utils/AttributeDefinitionUpdateActionUtils.java
@@ -19,36 +19,35 @@ import java.util.Optional;
 
 import static com.commercetools.sync.commons.utils.CommonTypeUpdateActionUtils.buildUpdateAction;
 import static com.commercetools.sync.commons.utils.OptionalUtils.filterEmptyOptionals;
-import static com.commercetools.sync.producttypes.utils.ProductTypeUpdateLocalizedEnumActionUtils.buildLocalizedEnumValuesUpdateActions;
-import static com.commercetools.sync.producttypes.utils.ProductTypeUpdatePlainEnumActionUtils.buildEnumValuesUpdateActions;
+import static com.commercetools.sync.producttypes.utils.LocalizedEnumsUpdateActionUtils.buildLocalizedEnumValuesUpdateActions;
+import static com.commercetools.sync.producttypes.utils.PlainEnumsUpdateActionUtils.buildEnumValuesUpdateActions;
 
-
-public final class AttributeDefinitionUpdateActionUtils {
+/**
+ * This class is only meant for the internal use of the commercetools-sync-java library.
+ */
+final class AttributeDefinitionUpdateActionUtils {
     /**
      * Compares all the fields of an {@link AttributeDefinition} and an {@link AttributeDefinitionDraft} and returns
      * a list of {@link UpdateAction}&lt;{@link ProductType}&gt; as a result. If both the {@link AttributeDefinition}
      * and the {@link AttributeDefinitionDraft} have identical fields, then no update action is needed and hence an
      * empty {@link List} is returned.
      *
-     * @param oldAttributeDefinition      the attribute definition which should be updated.
-     * @param newAttributeDefinitionDraft the attribute definition draft where we get the new fields.
+     * @param oldAttributeDefinition      the old attribute definition which should be updated.
+     * @param newAttributeDefinitionDraft the new attribute definition draft where we get the new fields.
      * @return A list with the update actions or an empty list if the attribute definition fields are identical.
      */
     @Nonnull
-    public static List<UpdateAction<ProductType>> buildActions(
+    static List<UpdateAction<ProductType>> buildActions(
         @Nonnull final AttributeDefinition oldAttributeDefinition,
         @Nonnull final AttributeDefinitionDraft newAttributeDefinitionDraft) {
 
-        final List<UpdateAction<ProductType>> updateActions;
-
-        updateActions = filterEmptyOptionals(
+        final List<UpdateAction<ProductType>> updateActions = filterEmptyOptionals(
             buildChangeLabelUpdateAction(oldAttributeDefinition, newAttributeDefinitionDraft),
             buildSetInputTipUpdateAction(oldAttributeDefinition, newAttributeDefinitionDraft),
             buildChangeIsSearchableUpdateAction(oldAttributeDefinition, newAttributeDefinitionDraft),
             buildChangeInputHintUpdateAction(oldAttributeDefinition, newAttributeDefinitionDraft),
             buildChangeAttributeConstraintUpdateAction(oldAttributeDefinition, newAttributeDefinitionDraft)
         );
-
 
         if (isPlainEnumAttribute(oldAttributeDefinition)) {
             updateActions.addAll(buildEnumValuesUpdateActions(
@@ -98,7 +97,7 @@ public final class AttributeDefinitionUpdateActionUtils {
      * @return A filled optional with the update action or an empty optional if the labels are identical.
      */
     @Nonnull
-    public static Optional<UpdateAction<ProductType>> buildChangeLabelUpdateAction(
+    static Optional<UpdateAction<ProductType>> buildChangeLabelUpdateAction(
         @Nonnull final AttributeDefinition oldAttributeDefinition,
         @Nonnull final AttributeDefinitionDraft newAttributeDefinition) {
 
@@ -119,7 +118,7 @@ public final class AttributeDefinitionUpdateActionUtils {
      * @return A filled optional with the update action or an empty optional if the labels are identical.
      */
     @Nonnull
-    public static Optional<UpdateAction<ProductType>> buildSetInputTipUpdateAction(
+    static Optional<UpdateAction<ProductType>> buildSetInputTipUpdateAction(
         @Nonnull final AttributeDefinition oldAttributeDefinition,
         @Nonnull final AttributeDefinitionDraft newAttributeDefinition) {
 
@@ -139,7 +138,7 @@ public final class AttributeDefinitionUpdateActionUtils {
      * @return A filled optional with the update action or an empty optional if the 'isSearchable' fields are identical.
      */
     @Nonnull
-    public static Optional<UpdateAction<ProductType>> buildChangeIsSearchableUpdateAction(
+    static Optional<UpdateAction<ProductType>> buildChangeIsSearchableUpdateAction(
         @Nonnull final AttributeDefinition oldAttributeDefinition,
         @Nonnull final AttributeDefinitionDraft newAttributeDefinition) {
 
@@ -159,7 +158,7 @@ public final class AttributeDefinitionUpdateActionUtils {
      * @return A filled optional with the update action or an empty optional if the input hints are identical.
      */
     @Nonnull
-    public static Optional<UpdateAction<ProductType>> buildChangeInputHintUpdateAction(
+    static Optional<UpdateAction<ProductType>> buildChangeInputHintUpdateAction(
         @Nonnull final AttributeDefinition oldAttributeDefinition,
         @Nonnull final AttributeDefinitionDraft newAttributeDefinition) {
 
@@ -179,7 +178,7 @@ public final class AttributeDefinitionUpdateActionUtils {
      * @return A filled optional with the update action or an empty optional if the attribute constraints are identical.
      */
     @Nonnull
-    public static Optional<UpdateAction<ProductType>> buildChangeAttributeConstraintUpdateAction(
+    static Optional<UpdateAction<ProductType>> buildChangeAttributeConstraintUpdateAction(
         @Nonnull final AttributeDefinition oldAttributeDefinition,
         @Nonnull final AttributeDefinitionDraft newAttributeDefinition) {