+++ src/test/java/com/commercetools/sync/producttypes/utils/producttypeactionutils/BuildLocalizedEnumUpdateActionsTest.java
@@ -4,9 +4,9 @@ import com.commercetools.sync.commons.exceptions.DuplicateKeyException;
 import io.sphere.sdk.commands.UpdateAction;
 import io.sphere.sdk.models.LocalizedEnumValue;
 import io.sphere.sdk.producttypes.ProductType;
-import io.sphere.sdk.producttypes.commands.updateactions.RemoveEnumValues;
 import io.sphere.sdk.producttypes.commands.updateactions.AddLocalizedEnumValue;
 import io.sphere.sdk.producttypes.commands.updateactions.ChangeLocalizedEnumValueOrder;
+import io.sphere.sdk.producttypes.commands.updateactions.RemoveEnumValues;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -14,10 +14,11 @@ import org.junit.rules.ExpectedException;
 import java.util.Collections;
 import java.util.List;
 
-import static com.commercetools.sync.producttypes.utils.ProductTypeUpdateLocalizedEnumActionUtils.buildLocalizedEnumValuesUpdateActions;
-import static java.util.Arrays.asList;
-import static org.assertj.core.api.Assertions.assertThat;
+import static com.commercetools.sync.producttypes.utils.LocalizedEnumsUpdateActionUtils.buildLocalizedEnumValuesUpdateActions;
 import static io.sphere.sdk.models.LocalizedString.ofEnglish;
+import static java.util.Arrays.asList;
+import static java.util.Collections.singletonList;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class BuildLocalizedEnumUpdateActionsTest {
     private static final LocalizedEnumValue ENUM_VALUE_A = LocalizedEnumValue.of("a", ofEnglish("label_a"));
@@ -142,7 +143,7 @@ public class BuildLocalizedEnumUpdateActionsTest {
         );
 
         assertThat(updateActions).containsExactly(
-            RemoveEnumValues.of("attribute_definition_name_1", asList("c"))
+            RemoveEnumValues.of("attribute_definition_name_1", singletonList("c"))
         );
     }
 
@@ -168,7 +169,7 @@ public class BuildLocalizedEnumUpdateActionsTest {
         );
 
         assertThat(updateActions).containsExactly(
-            RemoveEnumValues.of("attribute_definition_name_1", asList("c")),
+            RemoveEnumValues.of("attribute_definition_name_1", singletonList("c")),
             AddLocalizedEnumValue.of("attribute_definition_name_1", ENUM_VALUE_D)
         );
     }
@@ -199,7 +200,7 @@ public class BuildLocalizedEnumUpdateActionsTest {
         );
 
         assertThat(updateActions).containsExactly(
-            RemoveEnumValues.of("attribute_definition_name_1", asList("a")),
+            RemoveEnumValues.of("attribute_definition_name_1", singletonList("a")),
             ChangeLocalizedEnumValueOrder.of("attribute_definition_name_1", asList(
                 ENUM_VALUE_C,
                 ENUM_VALUE_B
@@ -254,7 +255,7 @@ public class BuildLocalizedEnumUpdateActionsTest {
         );
 
         assertThat(updateActions).containsExactly(
-            RemoveEnumValues.of("attribute_definition_name_1", asList("a")),
+            RemoveEnumValues.of("attribute_definition_name_1", singletonList("a")),
             AddLocalizedEnumValue.of("attribute_definition_name_1", ENUM_VALUE_D),
             ChangeLocalizedEnumValueOrder.of("attribute_definition_name_1", asList(
                 ENUM_VALUE_C,